Fixed Site Layout, Improved Nav Bar, Prepped Pages
Layout was a little busted so I made sure everything was flex and screen-based. Also decided to convert from px to rem while it was still easy. I think this is important for a text-based site such as this one. Added dark mode check on page load to prevent flash. Fixed up Nav bar, still a lot of work to do there but I'm going to focus on over the top user experience when I have the site in a more complete state in terms of content. Added titles to the pages.