Improved Nav, Added JM Logo, Layout
Added navigation buttons to the bottom of each page that can take you left or right. Will have to figure out how that works with the projects page since that one will also have redirect buttons for each project but thats a future problem!

Added JM logo as the home button on the nav bar and as the site favicon.

Improved layout with the content class which will keep everything centered

Theres going to be a lot of media queries with this site!